be/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.8.3
gunicorn==22.0.0

